backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Akbeeh <Akbeeeh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
apache-airflow = "^2.7.2"
pytest = "^7.4.3"
beautifulsoup4 = "^4.12.2"
boto3 = "^1.28.73"
werkzeug = "2.2.3"
connexion = "^2.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
